git pull code

+++ src/main/java/com/yimei/finance/controllers/admin/tpl/SwaggerDemoController.java
@@ -32,7 +32,7 @@ public class SwaggerDemoController {
     }
 
     @ApiOperation(value="获取用户详细信息", notes="根据url的id来获取用户详细信息")
-    @ApiImplicitParam(name = "id", value = "用户ID", required = true, dataType = "Long")
+    @ApiImplicitParam(name = "id", value = "用户ID", required = true, dataType = "int")
     @RequestMapping(value="/{id}", method=RequestMethod.GET)
     public UserTest getUser(@PathVariable Long id) {
         return users.get(id);
@@ -40,7 +40,7 @@ public class SwaggerDemoController {
 
     @ApiOperation(value="更新用户详细信息", notes="根据url的id来指定更新对象，并根据传过来的user信息来更新用户详细信息")
     @ApiImplicitParams({
-            @ApiImplicitParam(name = "id", value = "用户ID", required = true, dataType = "Long"),
+            @ApiImplicitParam(name = "id", value = "用户ID", required = true, dataType = "int"),
             @ApiImplicitParam(name = "user", value = "用户详细实体user", required = true, dataType = "User")
     })
     @RequestMapping(value="/{id}", method=RequestMethod.PUT)
@@ -53,7 +53,7 @@ public class SwaggerDemoController {
     }
 
     @ApiOperation(value="删除用户", notes="根据url的id来指定删除对象")
-    @ApiImplicitParam(name = "id", value = "用户ID", required = true, dataType = "Long")
+    @ApiImplicitParam(name = "id", value = "用户ID", required = true, dataType = "int")
     @RequestMapping(value="/{id}", method=RequestMethod.DELETE)
     public String deleteUser(@PathVariable Long id) {
         users.remove(id);
